[APPS] WBZ451 Application RTP Release - app_trps.c MISRA Check WSCH-185

+++ apps/ble/advanced_applications/ble_throughput/firmware/src/app_trps.c
@@ -427,7 +427,8 @@ void APP_TRPS_SendUpConnParaStatusToClient(uint16_t connHandle, uint8_t upConnPa
 
 void APP_TRPS_TxBufValidEvtProc(void) {
     uint16_t status = APP_RES_SUCCESS;
-    uint8_t grpId[] = {(uint8_t)(TRP_GRPID_NULL), (uint8_t)(TRP_GRPID_CHECK_SUM), (uint8_t)(TRP_GRPID_LOOPBACK), (uint8_t)(TRP_GRPID_FIX_PATTERN)};
+    uint8_t grpId[] = {TRP_GRPID_NULL, TRP_GRPID_CHECK_SUM, TRP_GRPID_LOOPBACK,
+		TRP_GRPID_FIX_PATTERN};
 
     if ((s_trpsConnList_t.workMode == TRP_WMODE_LOOPBACK) && (s_trpsConnList_t.leCircQueue_t.usedNum > 0)) {
         s_trpsConnList_t.srcType = APP_TRP_SRC_TYPE_LE;
